Move classes to the using package
