+++ testutil-server/src/main/java/io/spine/testing/server/blackbox/VerifyingCounter.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.testing.server;
+package io.spine.testing.server.blackbox;
 
 import com.google.common.annotations.VisibleForTesting;
