+++ testutil-server/src/test/java/io/spine/testing/server/blackbox/SubscriptionObserverTest.java
@@ -18,15 +18,11 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.testing.server;
+package io.spine.testing.server.blackbox;
 
 import com.google.common.testing.NullPointerTester;
 import com.google.common.truth.extensions.proto.ProtoTruth;
-import com.google.protobuf.Empty;
-import io.grpc.stub.StreamObserver;
 import io.spine.client.SubscriptionUpdate;
-import io.spine.core.Response;
-import io.spine.core.Status;
 import io.spine.testing.server.given.GivenSubscriptionUpdate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
